+++ src/api/handler/dto/transaction.rs
@@ -1,0 +1,37 @@
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+use crate::models::transaction::TransactionModel;
+
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
+pub struct TransactionDto {
+    pub tx_hash: String,
+    pub unsigned: serde_json::Value,
+    pub script_execution_ok: bool,
+    pub contract_inputs: serde_json::Value,
+    pub generated_outputs: serde_json::Value,
+    pub input_signatures: Vec<Option<String>>,
+    pub script_signatures: Vec<Option<String>>,
+    #[schema(example = "2023-01-01T00:00:00", nullable = true)]
+    pub created_at: Option<String>,
+    #[schema(example = "2023-01-01T00:00:00", nullable = true)]
+    pub updated_at: Option<String>,
+    pub block_hash: String,
+}
+
+impl From<TransactionModel> for TransactionDto {
+    fn from(model: TransactionModel) -> Self {
+        Self {
+            tx_hash: model.tx_hash,
+            unsigned: model.unsigned,
+            script_execution_ok: model.script_execution_ok,
+            contract_inputs: model.contract_inputs,
+            generated_outputs: model.generated_outputs,
+            input_signatures: model.input_signatures,
+            script_signatures: model.script_signatures,
+            created_at: model.created_at.map(|dt| dt.format("%Y-%m-%dT%H:%M:%S").to_string()),
+            updated_at: model.updated_at.map(|dt| dt.format("%Y-%m-%dT%H:%M:%S").to_string()),
+            block_hash: model.block_hash.to_string(),
+        }
+    }
+}
