+++ src/models/block.rs
@@ -1,7 +1,6 @@
 use chrono::NaiveDateTime;
 use diesel::prelude::*;
 use serde::Serialize;
-use utoipa::ToSchema;
 
 use crate::types::{BlockHash, DEFAULT_GROUP_NUM};
