+++ src/models/transaction.rs
@@ -1,10 +1,8 @@
+use crate::models::BlockModel;
+use crate::{schema::transactions, types::BlockHash};
 use chrono::NaiveDateTime;
 use diesel::prelude::*;
 use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
-
-use crate::models::BlockModel;
-use crate::{schema::transactions, types::BlockHash};
 #[derive(
     Queryable,
     Selectable,
